# swagger.yaml
openapi: 3.0.3
info:
  title: Proyecto Final Programacion Backend CoderHouse
  description: Este es el Proyecto Final del curso de Programacion Backend de CoderHouse. 
    Se trata de una aplicacion backend de e-commerce con la posibilidad de obtener los productos disponibles, generar usuarios con sus carritos de compra y gestionarlos, generar pedidos y enviar mensajes automaticos.
  version: 1.0.0
components:
  schemas:
...
    Pedidos:
      type: object
      required:
        - _id
        - usuario
        - productos
        - direccionEntrega
        - estado
      properties:
        _id:
          type: string
          description: id del pedido
        usuario:
          type: object
          description: datos de usuario que hizo el pedido
        productos:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              nombre:
                type: string
              cantidad:
                type: number
              precio:
                type: number
          description: lista de productos del pedido
        direccionEntrega:
          type: string
          description: direcicon de entrega del pedido
        estado:
          type: string
          description: estado del pedido
      example:
        _id: "63100d958340e977a004f3a5"
        usuario: "630f6a3fc3f12bf3614f9819"
        productos:
          - _id: "630f71d273bd86d8c40db5fc"
            nombre: "Resident Evil Village"
            cantidad: 15
            precio: 1300
          - _id: "630f701273bd86d8c40db5d7"
            nombre: "Red Dead Redemption 2"
            cantidad: 10
            precio: 2500
        direccionEntrega: "Av. Rivadavia 4319"
        estado: Creado
...
    Error:
      type: object
      properties:
        message:
          type: string
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
servers:
  - url: http://localhost:8080/api
    description: Server Local
  - url: https://tp-final-backend.herokuapp.com
    description: Server de Produccion
paths:
...
  /pedido:
    get:
      summary: devuelve todas las ordenes
      tags:
        - Pedidos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  pedido:
                    $ref: '#/components/schemas/Pedidos'
                example:
                  - estado: Creado
                    _id: 630f72e973bd86d8c40db62b
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f701273bd86d8c40db5d7
                        cantidad: 3
                      - _id: 630f71c973bd86d8c40db5f6
                        cantidad: 2
                    createdAt: '2022-08-31T14:40:41.520Z'
                    updatedAt: '2022-08-31T14:40:41.520Z'
                  - estado: Creado
                    _id: 630f8a82f4e745bf9bce2d28
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f71c973bd86d8c40db5f6
                        cantidad: 2
                        precio: 2500
                    createdAt: '2022-08-31T16:21:22.633Z'
                    updatedAt: '2022-08-31T16:21:22.633Z'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '403':
          description: El usuario no tiene permisos para acceder al recurso
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Forbidden' 
              example:
                message: "No tiene permisos para acceder al recurso"
    post:
      summary: Crea un pedido a partir del carrito del usuario
      tags:
        - Pedidos
      parameters:
        - name: user.carrito
          in: cookie
          required: true
          description: ID del carrito del usuario logueado.
          schema:
            type : string
          example: "630f6a3fc3f12bf3614f981b"
      requestBody:
          content:
            application/json:
              schema:
                properties:
                  direccion:
                    type: string
                example:
                  direccion: "Av. Rivadaiva 4319"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message: 
                    type: string
                  carrito:
                    $ref: '#/components/schemas/Pedidos'
                example:
                  message: Pedido creado
                  pedido:
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f701273bd86d8c40db5d7
                        nombre: Red Dead Redemption 20
                        cantidad: 22
                        precio: 2500
                      - _id: 630f71d273bd86d8c40db5fc
                        nombre: Red Dead 4
                        cantidad: 5
                        precio: 2500
                    direccionEntrega: 'Av. Rivadavia 4319'
                    estado: Creado
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '400':
          description: "Carrito vacio"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "No se puede completar el pedido. El carrito esta vacio"
  /pedido/myorders:
    get:
      summary: devuelve todas las ordenes del usuario
      tags:
        - Pedidos
      parameters:
        - name: user._id
          in: cookie
          required: true
          description: ID del Usuario
          schema:
            type : string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  pedido:
                    $ref: '#/components/schemas/Pedidos'
                example:
                  - _id: 630f72e973bd86d8c40db62b
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f701273bd86d8c40db5d7
                        nombre: 'Resident Evil'
                        cantidad: 3
                        precio: 1500
                      - _id: 630f71c973bd86d8c40db5f6
                        nombre: 'GTA'
                        cantidad: 2
                        precio: 2500
                    direccionEntrega: "Av. San Juan 3115"
                    estado: Creado
                    createdAt: '2022-08-31T14:40:41.520Z'
                    updatedAt: '2022-08-31T14:40:41.520Z'
                  - _id: 630f8a82f4e745bf9bce2d28
                    usuario: 630f6a3fc3f12bf3614f9819
                    productos:
                      - _id: 630f71c973bd86d8c40db5f6
                        nombre: "GTA"
                        cantidad: 2
                        precio: 2500
                    direccionEntrega: "Av. Rivadavia 4319"
                    estado: Creado
                    createdAt: '2022-08-31T16:21:22.633Z'
                    updatedAt: '2022-08-31T16:21:22.633Z'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
  /pedido/last:
    get:
      summary: devuelve la ultima orden del usuario
      tags:
        - Pedidos
      parameters:
        - name: user._id
          in: cookie
          required: true
          description: ID del Usuario
          schema:
            type : string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  pedido:
                    $ref: '#/components/schemas/Pedidos'
                example:
                  _id: 630f8a82f4e745bf9bce2d28
                  usuario: 630f6a3fc3f12bf3614f9819
                  productos:
                    - _id: 630f71c973bd86d8c40db5f6
                      nombre: "GTA"
                      cantidad: 2
                      precio: 2500
                  direccionEntrega: "Av. Rivadavia 4319"
                  estado: Creado
                  createdAt: '2022-08-31T16:21:22.633Z'
                  updatedAt: '2022-08-31T16:21:22.633Z'
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
  /pedido/{id}/productos:
    get:
      summary: Devuelve el detalle de productos del Pedido con el id indicado
      tags:
        - Pedidos
      parameters:
        - name: id
          in: path
          required: true
          description: ID del Pedido
          schema:
            type : string
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                productos:
                  - _id: 630f71c973bd86d8c40db5f6
                    nombre: "GTA"
                    cantidad: 2
                    precio: 2500
                  - _id: 630f701273bd86d8c40db5d7
                    nombre: 'Resident Evil'
                    cantidad: 3
                    precio: 1500                  
        '401':
          description: El usuario no esta logueado/autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
              example:
                message: "Usuario no autenticado"
        '403':
          description: El usuario no tiene permisos para acceder al recurso
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Forbidden' 
              example:
                message: "No tiene permisos para acceder al recurso"
        '404':
          description: "Pedido no encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
              example:
                message: "Pedido encontrado"